Module Pbrt

Protobuf runtime library

type payload_kind =

| Varint

| Bits32

| Bits64

| Bytes

Type of payload in a given field.

This is only the wire type, the generated code will have a more precise type in general.

module Decoder : sig ... end

Decoding protobufs.

module Encoder : sig ... end

Encoding protobufs.

module Repeated_field : sig ... end

Optimized representation for repeated fields

module Pp : sig ... end

Runtime functions for Pretty Printing functionality