Module Pbrt_options

Protobuf

module E : sig ... end

All exceptions which could be raised by the generated option parsers.

type constant =

| Constant_string of string

| Constant_bool of bool

| Constant_int of int

| Constant_float of float

| Constant_literal of string

type message_literal = (string * value) list

and list_literal = value list

and value =

| Scalar_value of constant

| Message_literal of message_literal

| List_literal of list_literal

Helper module for the generated code for common functionality

val string : value -> string -> string -> string

val float : value -> string -> string -> float

val int32 : value -> string -> string -> int32

val int64 : value -> string -> string -> int64

val int : value -> string -> string -> int

val bool : value -> string -> string -> bool

val bytes : value -> string -> string -> bytes

val unit : value -> string -> string -> unit